Package: rename package name for eclipse
