+++ app/build.gradle
@@ -14,7 +14,7 @@ android {
     buildToolsVersion "29.0.3"
 
     defaultConfig {
-        applicationId "org.cna.keyple.demo.control"
+        applicationId "org.eclipse.keyple.demo.control"
         minSdkVersion 21
         targetSdkVersion 29
         versionCode versionMajor * 1000000 + versionMinor * 10000 + versionPatch * 100 + versionBeta
