+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.cna.keyple.demo.control">
+    package="org.eclipse.keyple.demo.control">
 
     <application
         android:allowBackup="true"
@@ -11,23 +11,23 @@
         android:theme="@style/Theme.KeypleControl">
 
         <activity
-            android:name=".activities.NetworkInvalidActivity"
+            android:name="org.eclipse.keyple.demo.control.activities.NetworkInvalidActivity"
             android:label="@string/title_activity_network_invalid"
             android:theme="@style/Theme.KeypleControl.NoActionBar"/>
         <activity
-            android:name=".activities.CardContentActivity"
+            android:name="org.eclipse.keyple.demo.control.activities.CardContentActivity"
             android:label="@string/title_activity_card_content"
             android:theme="@style/Theme.KeypleControl.NoActionBar"/>
         <activity
-            android:name=".activities.CardReaderActivity"
+            android:name="org.eclipse.keyple.demo.control.activities.CardReaderActivity"
             android:label="@string/title_activity_card_reader"
             android:theme="@style/Theme.KeypleControl.NoActionBar" />
         <activity
-            android:name=".activities.HomeActivity"
+            android:name="org.eclipse.keyple.demo.control.activities.HomeActivity"
             android:label="@string/title_activity_home"
             android:theme="@style/Theme.KeypleControl.NoActionBar" />
         <activity
-            android:name=".activities.SettingsActivity"
+            android:name="org.eclipse.keyple.demo.control.activities.SettingsActivity"
             android:label="@string/app_name"
             android:theme="@style/Theme.KeypleControl.NoActionBar">
             <intent-filter>
