+++ app/src/main/java/org/eclipse/keyple/demo/control/activities/CardReaderActivity.kt
@@ -10,16 +10,16 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.cna.keyple.demo.control.activities
+package org.eclipse.keyple.demo.control.activities
 
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
-import org.cna.keyple.demo.control.R
-import org.cna.keyple.demo.control.models.CardReaderResponse
-import org.cna.keyple.demo.control.models.CardTitle
-import org.cna.keyple.demo.control.models.Validation
+import org.eclipse.keyple.demo.control.R
+import org.eclipse.keyple.demo.control.models.CardReaderResponse
+import org.eclipse.keyple.demo.control.models.CardTitle
+import org.eclipse.keyple.demo.control.models.Validation
 import java.text.SimpleDateFormat
 import java.util.*
