+++ app/src/main/java/org/eclipse/keyple/demo/control/activities/HomeActivity.kt
@@ -10,13 +10,13 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.cna.keyple.demo.control.activities
+package org.eclipse.keyple.demo.control.activities
 
 import android.content.Intent
 import android.os.Bundle
 import android.widget.ImageButton
 import androidx.appcompat.app.AppCompatActivity
-import org.cna.keyple.demo.control.R
+import org.eclipse.keyple.demo.control.R
 
 class HomeActivity : AppCompatActivity() {
