+++ app/src/main/java/org/eclipse/keyple/demo/control/adapters/ValidationsRecyclerAdapter.kt
@@ -10,15 +10,15 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.cna.keyple.demo.control.adapters
+package org.eclipse.keyple.demo.control.adapters
 
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import kotlinx.android.synthetic.main.validation_recycler_row.view.*
-import org.cna.keyple.demo.control.R
-import org.cna.keyple.demo.control.inflate
-import org.cna.keyple.demo.control.models.Validation
+import org.eclipse.keyple.demo.control.R
+import org.eclipse.keyple.demo.control.inflate
+import org.eclipse.keyple.demo.control.models.Validation
 import java.text.SimpleDateFormat
 import java.util.*
