+++ app/src/main/java/org/eclipse/keyple/demo/control/models/CardReaderResponse.kt
@@ -9,15 +9,14 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.cna.keyple.demo.control.models
+package org.eclipse.keyple.demo.control.models
 
-enum class Status(val status: String) {
-    LOADING("loading"),
-    ERROR("error"),
-    TICKETS_FOUND("tickets_found"),
-    INVALID_CARD("invalid_card"),
-    EMPTY_CARD("empty_card"),
-    WRONG_CARD("wrong_card"),
-    DEVICE_CONNECTED("device_connected"),
-    SUCCESS("success");
-}
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class CardReaderResponse(
+    val cardType: String,
+    val lastValidationsList: ArrayList<Validation>,
+    val titlesList: ArrayList<CardTitle>
+): Parcelable
