+++ app/src/main/res/layout/activity_card_reader.xml
@@ -17,7 +17,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.cna.keyple.demo.control.activities.CardReaderActivity">
+    tools:context="org.eclipse.keyple.demo.control.activities.CardReaderActivity">
 
     <include
         layout="@layout/logo_toolbar"
