+++ app/src/main/res/layout/activity_home.xml
@@ -17,7 +17,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.cna.keyple.demo.control.activities.HomeActivity">
+    tools:context="org.eclipse.keyple.demo.control.activities.HomeActivity">
 
     <include
         layout="@layout/logo_toolbar"
